chore: CarRandomMoveCondtion Fake 클래스를 test 패키지로 이동
